pyproject.toml
[project]
name = "django-tweet"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.96.0",
    "dj-database-url>=3.1.2",
    "django>=6.0.4",
    "gunicorn>=25.3.0",
    "openai>=2.32.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "whitenoise>=6.12.0",
]
